To remove compile configuration from all tests 2/2
