+++ src/test/resources/app_builds_resolved_configurations-01.json
@@ -3,30 +3,16 @@
     "configuration": {
       "allDependencies": [
         "DefaultExternalModuleDependency{group='org.scala-lang', name='scala-library', version='2.11.12', configuration='default'}",
-        "DefaultExternalModuleDependency{group='org.scala-lang', name='scala-reflect', version='2.12.8', configuration='default'}",
-        "DefaultExternalModuleDependency{group='org.scalaz', name='scalaz-core_@dsv@', version='7.2.28', configuration='default'}",
-        "DefaultExternalModuleDependency{group='org.scala-lang', name='scala-library', version='@dsv@.+', configuration='default'}"
+        "DefaultExternalModuleDependency{group='org.scalaz', name='scalaz-core_@dsv@', version='7.2.28', configuration='default'}"
       ],
-      "name": "crossBuildSpark230_211Compile"
+      "name": "crossBuildSpark230_211Api"
     },
     "sourceSet": "crossBuildSpark230_211",
     "resolvedConfiguration": {
       "hasError": false,
-      "resolvedArtifacts": [
-        "scala-reflect@_2_11_12_@.jar (org.scala-lang:scala-reflect:2.11.12)",
-        "scalaz-core_2.11@_7_2_28_@.jar (org.scalaz:scalaz-core_2.11:7.2.28)",
-        "scala-library@_2_11_12_@.jar (org.scala-lang:scala-library:2.11.12)"
-      ],
-      "files": [
-        "scala-reflect-2.11.12.jar",
-        "scalaz-core_2.11-7.2.28.jar",
-        "scala-library-2.11.12.jar"
-      ],
-      "firstLevelModuleDependencies": [
-        "org.scala-lang:scala-reflect:2.11.12;@defaultOrRuntime@",
-        "org.scalaz:scalaz-core_2.11:7.2.28;@defaultOrRuntime@",
-        "org.scala-lang:scala-library:2.11.12;@defaultOrRuntime@"
-      ]
+      "resolvedArtifacts": null,
+      "files": null,
+      "firstLevelModuleDependencies": null
     }
   },
   {
@@ -35,9 +21,7 @@
         "DefaultProjectDependency{dependencyProject='project ':lib2'', configuration='crossBuildSpark230_211Producer'}",
         "DefaultProjectDependency{dependencyProject='project ':lib'', configuration='crossBuildSpark230_211Producer'}",
         "DefaultExternalModuleDependency{group='org.scala-lang', name='scala-library', version='2.11.12', configuration='default'}",
-        "DefaultExternalModuleDependency{group='org.scala-lang', name='scala-reflect', version='2.12.8', configuration='default'}",
-        "DefaultExternalModuleDependency{group='org.scalaz', name='scalaz-core_@dsv@', version='7.2.28', configuration='default'}",
-        "DefaultExternalModuleDependency{group='org.scala-lang', name='scala-library', version='@dsv@.+', configuration='default'}"
+        "DefaultExternalModuleDependency{group='org.scalaz', name='scalaz-core_@dsv@', version='7.2.28', configuration='default'}"
       ],
       "name": "crossBuildSpark230_211Implementation"
     },
@@ -55,9 +39,7 @@
         "DefaultProjectDependency{dependencyProject='project ':lib2'', configuration='crossBuildSpark230_211Producer'}",
         "DefaultProjectDependency{dependencyProject='project ':lib'', configuration='crossBuildSpark230_211Producer'}",
         "DefaultExternalModuleDependency{group='org.scala-lang', name='scala-library', version='2.11.12', configuration='default'}",
-        "DefaultExternalModuleDependency{group='org.scala-lang', name='scala-reflect', version='2.12.8', configuration='default'}",
-        "DefaultExternalModuleDependency{group='org.scalaz', name='scalaz-core_@dsv@', version='7.2.28', configuration='default'}",
-        "DefaultExternalModuleDependency{group='org.scala-lang', name='scala-library', version='@dsv@.+', configuration='default'}"
+        "DefaultExternalModuleDependency{group='org.scalaz', name='scalaz-core_@dsv@', version='7.2.28', configuration='default'}"
       ],
       "name": "crossBuildSpark230_211CompileClasspath"
     },
@@ -65,21 +47,18 @@
     "resolvedConfiguration": {
       "hasError": false,
       "resolvedArtifacts": [
-        "scala-reflect@_2_11_12_@.jar (org.scala-lang:scala-reflect:2.11.12)",
         "scalaz-core_2.11@_7_2_28_@.jar (org.scalaz:scalaz-core_2.11:7.2.28)",
         "scala-library@_2_11_12_@.jar (org.scala-lang:scala-library:2.11.12)",
         "lib2_2.11.jar (project :lib2)",
         "lib_2.11.jar (project :lib)"
       ],
       "files": [
-        "scala-reflect-2.11.12.jar",
         "scalaz-core_2.11-7.2.28.jar",
         "scala-library-2.11.12.jar",
         "lib2_2.11-1.0-SNAPSHOT.jar",
         "lib_2.11-1.0-SNAPSHOT.jar"
       ],
       "firstLevelModuleDependencies": [
-        "org.scala-lang:scala-reflect:2.11.12;@defaultOrCompile@",
         "org.scalaz:scalaz-core_2.11:7.2.28;@defaultOrCompile@",
         "org.scala-lang:scala-library:2.11.12;@defaultOrCompile@",
         "com.github.prokod.it:lib2:1.0-SNAPSHOT;crossBuildSpark230_211Producer",
@@ -91,9 +70,7 @@
     "configuration": {
       "allDependencies": [
         "DefaultExternalModuleDependency{group='org.scala-lang', name='scala-library', version='2.11.12', configuration='default'}",
-        "DefaultExternalModuleDependency{group='org.scala-lang', name='scala-reflect', version='2.12.8', configuration='default'}",
         "DefaultExternalModuleDependency{group='org.scalaz', name='scalaz-core_@dsv@', version='7.2.28', configuration='default'}",
-        "DefaultExternalModuleDependency{group='org.scala-lang', name='scala-library', version='@dsv@.+', configuration='default'}",
         "DefaultProjectDependency{dependencyProject='project ':lib2'', configuration='crossBuildSpark230_211Producer'}",
         "DefaultProjectDependency{dependencyProject='project ':lib'', configuration='crossBuildSpark230_211Producer'}"
       ],
@@ -103,21 +80,18 @@
     "resolvedConfiguration": {
       "hasError": false,
       "resolvedArtifacts": [
-        "scala-reflect@_2_11_12_@.jar (org.scala-lang:scala-reflect:2.11.12)",
         "scalaz-core_2.11@_7_2_28_@.jar (org.scalaz:scalaz-core_2.11:7.2.28)",
         "scala-library@_2_11_12_@.jar (org.scala-lang:scala-library:2.11.12)",
         "lib2_2.11.jar (project :lib2)",
         "lib_2.11.jar (project :lib)"
       ],
       "files": [
-        "scala-reflect-2.11.12.jar",
         "scalaz-core_2.11-7.2.28.jar",
         "scala-library-2.11.12.jar",
         "lib2_2.11-1.0-SNAPSHOT.jar",
         "lib_2.11-1.0-SNAPSHOT.jar"
       ],
       "firstLevelModuleDependencies": [
-        "org.scala-lang:scala-reflect:2.11.12;@defaultOrRuntime@",
         "org.scalaz:scalaz-core_2.11:7.2.28;@defaultOrRuntime@",
         "org.scala-lang:scala-library:2.11.12;@defaultOrRuntime@",
         "com.github.prokod.it:lib2:1.0-SNAPSHOT;crossBuildSpark230_211Producer",
